[log] Minor touchups to the structure and the abstractions
