+++ .gitignore
@@ -8,3 +8,7 @@ cargo-timing*.html
 # often need a place to throw one-off files
 /tmp
 /temp
+
+# Log files
+log.*
+!log.rs
