+++ bve/src/log/data.rs
@@ -1,31 +1,7 @@
 pub use crate::log::common::Data as MessageValue;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use std::io::{Read, Write};
 use tracing_core::Level;
-
-#[derive(Clone, Debug)]
-pub enum SerializationMethod {
-    Bincode,
-    Json,
-    JsonPretty,
-}
-
-impl SerializationMethod {
-    pub fn serialize(&self, writer: &mut impl Write, message: &Message) {
-        match self {
-            Self::Bincode => bincode::serialize_into(writer, message).expect("Bincode serialization failed"),
-            Self::Json => {
-                serde_json::to_writer(&mut *writer, message).expect("Json serialization failed");
-                writeln!(writer, "");
-            }
-            Self::JsonPretty => {
-                serde_json::to_writer_pretty(&mut *writer, message).expect("Json serialization failed");
-                writeln!(writer, "");
-            }
-        }
-    }
-}
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub enum Message {
